+++ engine/wire.js
@@ -134,7 +134,7 @@ const NON_CARD_SIDE_FIELDS = [
   "id","name","hero","heroKey","hp","maxHp","int","baseInt","intWas",
   "res","ap","wasted","counters","weaponUsed","heroTapped","buffNext","buffQ","atkBuff","defMod","gaNext","gaNextQ","costOff","instantNextQ","defCapNext","defActionBuff","wardRider",
   "runeHitNext","amp","ward","awd","wardTurn","awdTurn","arcShield",
-  "lifeLock","namedBuff","dracNext","marked","fatigue","hist",
+  "lifeLock","namedBuff","dracNext","dracChain","marked","fatigue","hist",
   "blockH","blockG","blockRx","blockedHand","chainBlocked","intimidated","paySel",
   "nextTurn"
 ];